+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.9]
+
+### Updated
+
+- Refactoring of observability class
+- Refactoring of prompt manager class
+- Fix uvicorn setup
+
 ## [0.8.8]
 
 ### Fixed